# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nebula Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
  dev: 'dev',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  isBanned: boolean;
  /** @nullable */
  banReason?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  user: 'user',
  admin: 'admin',
  dev: 'dev',
} as const;

export interface UserInput {
  username: string;
  email: string;
  password: string;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  user: 'user',
  admin: 'admin',
  dev: 'dev',
} as const;

export interface UserUpdate {
  username?: string;
  email?: string;
  role?: UserUpdateRole;
  password?: string;
}

export interface BanInput {
  reason: string;
}

export type ServerStatus = typeof ServerStatus[keyof typeof ServerStatus];


export const ServerStatus = {
  running: 'running',
  stopped: 'stopped',
  installing: 'installing',
  suspended: 'suspended',
} as const;

export interface Server {
  id: number;
  name: string;
  status: ServerStatus;
  owner: string;
  node: string;
  ram: number;
  cpu: number;
  disk: number;
}

export interface ProtectFeature {
  id: number;
  name: string;
  description: string;
  enabled: boolean;
}

export interface ProtectToggleInput {
  id: number;
  enabled: boolean;
}

export interface MaintenanceStatus {
  enabled: boolean;
  message: string;
}

export interface MaintenanceInput {
  enabled: boolean;
  message?: string;
}

export interface OverviewStats {
  totalUsers: number;
  totalServers: number;
  activeServers: number;
  bannedUsers: number;
  adminCount: number;
  devCount: number;
}

export interface ActivityLog {
  id: number;
  action: string;
  user: string;
  timestamp: string;
  /** @nullable */
  details?: string | null;
}