# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nebula Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  action: string;
  user: string;
  timestamp: string;
  /** @nullable */
  details?: string | null;
}