# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nebula Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerStatus } from './serverStatus';

export interface Server {
  id: number;
  name: string;
  status: ServerStatus;
  owner: string;
  node: string;
  ram: number;
  cpu: number;
  disk: number;
}